Se arregló el problema de falta
